moved main list page layout to its own component
